# Used webpack plugins to compress js and css assets.