--- a/package.json
+++ b/package.json
@@ -37,6 +37,7 @@
     "sass-loader": "^4.0.0",
     "style-loader": "^0.13.1",
     "susy": "^2.2.12",
+    "uglify-js": "^2.8.21",
     "uglifyjs-webpack-plugin": "^0.3.1",
     "url-loader": "^0.5.7",
     "vue-hot-reload-api": "^1.3.2",