@cappadona Locomotive CMS (engine) should handle asset compression, so not sure if this is PR is needed.