@cappadona 2017-03-31
I almost forgot...I also ran numbers for this PR vs. having Locomotive (via Steam) minify and compress JS & CSS -- a new feature [introduced with Engine `v3.3`](https://locomotive-v3.readme.io/v3.3/docs/version-33#section-features-improvements) I'm including those here.

Again, aside from the anomaly with `main.css` mentioned above, with the limited data in hand there doesn't appear to be much of a benefit in using Locomotive's minifying instead of having Webpack handle it.

One thing worth noting is that since Locomotive doesn't minify while hosting via wagon, it doesn't slow down or affect the development workflow until you go to deploy assets to an engine (which takes significantly longer now). On the other hand, the Webpack plugins are quicker to minify but will run more frequently during the development process. Initial start time for `npm start` was increased but since Webpack is watching files for changes, I don't suspect the inclusion of minifying assets to reduce the developer experience to an extent it will become bothersome. Also, I'm sure we could configure webpack to only minify when prepping assets for production, although that would introduce changes to the existing workflow as well...okay, talk amongst yourselves.

![200w_d](https://cloud.githubusercontent.com/assets/409842/24555421/1860c4b0-15ff-11e7-8ca4-66f15e28ba06.gif)

### Performance of Engine `v3.3` vs. Baseline
> Percentage of decrease from [baseline data](https://github.com/cul-it/mann-wagon/issues/831) in bold

| Environment | `DOMContentLoaded` | Load |
| --- | --- | --- |
| local engine `v3.3.0.rc2` | 28.17s (**51.4%**) | 54.86s (**6.1%**) |

### Performance of Engine `v3.3` vs. PR #839 
> Percentage of difference from [PR #839 data](https://github.com/cul-it/mann-wagon/issues/838#issuecomment-290712601) in bold

| Environment | `DOMContentLoaded` | Load |
| --- | --- | --- |
| local engine `v3.3.0.rc2` | 28.17s (**3.4% INCREASE**) | 54.86s (**3.1% DECREASE**) |

> Only run against local engine since Locomotive [doesn't minify assets when serving via wagon](https://github.com/locomotivecms/wagon/commit/f3f75d7) and there was no significant difference between baseline numbers in local vs dev engines.

### Performance of Engine `v3.3` combined with PR #839 vs. PR #839 only
> Percentage of difference from [PR #839 data](https://github.com/cul-it/mann-wagon/issues/838#issuecomment-290712601) in bold

| Environment | `DOMContentLoaded` | Load |
| --- | --- | --- |
| wagon `v2.3.0rc1` | 24.76s (**0.09% DECREASE**) | 43.92s (**0.24% INCREASE**) |
| local engine `v3.3.0.rc2` | 29.02s (**6.5% INCREASE**) | 55.51s (**2% DECREASE**) |

### File size reduction

| Asset | Baseline | PR #839 | Loco 3.3 |
| --- | --- | --- | --- |
| Overall payload | 15.2MB | 12.8MB (**15.8%**) | 12.5MB (**17.8%**) |
| `homepageEvents.bundle.js` | 2.2MB | 764KB (**65.3%**) | 822KB (**62.6%**) |
| `main.css` | 722MB | 558KB (**22.7%**) | 211KB (**70.8%**) |
| `vendor.bundle.js` | 567KB | 350KB (**38.3%**) | 350KB (**38.3%**) |
| `experts.bundle.js` | 421KB | 156KB (**63%**) | 155KB (**63.2%**) |
| `spotlights.bundle.js` | 415KB | 153KB (**63.1%**) | 152KB (**63.4%**) |