@tap87 2017-04-03
@cappadona It seems like UglifyJS is pretty chatty. https://github.com/webpack/webpack/issues/2352.
I think we can safely ignore these, and configure uglify to not show the warnings. I looked at a few of these, some are simple issue like setting a variable to true/false when initializing in our code. Some are from external libraries that webpack bundles up.