@tap87 2017-04-03
@cappadona main.css in the browser has additional info about source mapping. There is an option in [cssnano to set it to true to print it inline](http://cssnano.co/options/).