@cappadona 2017-04-03
Good catch on the sourcemap for `main.css`. I'm seeing it in [production](http://mannlib.cornell.edu/sites/56e8640e3d262f7310000000/theme/stylesheets/main.css?1479159427), but not dev, which makes me think that cssnano is not the culprit but that something further upstream in the Webpack config is responsible.