@tap87 2017-04-03
Yes, that was my guess as well, probably the style loader.