@cappadona 2017-04-03
Interesting, looks like [default for `sourceMap` option in css-loader is false](https://github.com/webpack-contrib/css-loader#options-1) and I'm not seeing where we explicitly enabled it in our config.