@cappadona 2017-04-03
It might be the [`extract-text-webpack-plugin`](https://github.com/webpack-contrib/extract-text-webpack-plugin/blob/webpack-1/README.md)...need to take a closer look.