@cappadona 2017-04-03
You can see that in a previous incarnation of our Webpack config, I had enabled sourcemaps, but they should have been disabled with 43a9402 from July 2016.

Since I'm not currently seeing sourcemaps in the latest on the dev branch, I'm thinking we've finally left this ghost behind, but I'll be curious to hear what you find after rebasing your PR on dev.