@cappadona 2017-04-06
Found the [*source*](https://github.com/cul-it/mann-wagon/commit/89067444296857b98a8074447358b04571181940#diff-642224899a7aee531d6c26057b2b1a05R5) of our sourcemaps :)